Fixed Commonship
- fixed move method, now is for a singular entity;
- now implements Collidable  and extends Collider;
- minor fixes.
